terminal/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.0
google-genai
feedparser
duckduckgo-search
python-multipart
websockets
aiofiles
pydantic-settings
plotly
numpy
scipy
yfinance
ddgs
docker